finetuning/requirements.txt
# 파인튜닝을 위한 필수 패키지들
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
accelerate>=0.20.0
peft>=0.4.0
evaluate>=0.4.0

# 데이터 처리
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# 시각화
matplotlib>=3.5.0
seaborn>=0.11.0

# 평가 메트릭
rouge-score>=0.1.2
sacrebleu>=2.3.0

# 설정 파일 처리
PyYAML>=6.0

# 로깅 및 모니터링
tensorboard>=2.10.0
wandb>=0.15.0  # 선택사항

# GPU 최적화 (필수 - GGUF 호환 파인튜닝용)
bitsandbytes>=0.41.0  # 4bit/8bit 양자화용
# deepspeed>=0.9.0      # 대용량 모델용 (선택사항)

# 추가 유틸리티
tqdm>=4.64.0
colorama>=0.4.4
